Use double for metadata numbers

+++ src/main/java/com/box/sdk/BoxFolder.java
@@ -876,7 +876,7 @@ public class BoxFolder extends BoxItem implements Iterable<BoxItem.Info> {
                 for (JsonValue value : metadata.getOperations()) {
                     if (value.asObject().get("value").isNumber()) {
                         metadataToUpdate.add(value.asObject().get("path").asString(),
-                                value.asObject().get("value").asFloat());
+                                value.asObject().get("value").asDouble());
                     } else if (value.asObject().get("value").isString()) {
                         metadataToUpdate.add(value.asObject().get("path").asString(),
                                 value.asObject().get("value").asString());
